+++ client/components/DBDisplay/Canvas.tsx
@@ -1,5 +1,5 @@
 // React & React Router & React Query Modules;
-import React, { useCallback, useRef, useState } from "react";
+import React from "react";
 import { useMutation } from "react-query";
 
 //Components imported;
